+++ src/lib/shared/cache-keys.shared.ts
@@ -14,5 +14,9 @@ export const CacheKeys = {
     props.user,
     props.repo,
     props.number
-  ]
+  ],
+  markdownPreview: (props: {
+    repositoryPath: `${string}/${string}`;
+    content: string;
+  }) => ["MARKDOWN_PREVIEW", props.repositoryPath, props.content]
 } as const;
